Added admin checks for client endpoints

+++ src/routes/clients.js
@@ -1,3 +1,4 @@
+const { verifyTokenIsAdmin } = require('../util/auth')
 const Mongo = require('../db/index')
 
 /**
@@ -10,10 +11,15 @@ const main = server => {
   // Retrieve a list of all currently created Clients.
   server.get({ path: '/clients', version: '1' }, async (req, res, next) => {
     try {
-      const cService = Mongo.getClientService()
-      const clients = await cService.getClients()
+      const isAdmin = await verifyTokenIsAdmin(req.headers['request-token'])
+      if (isAdmin) {
+        const cService = Mongo.getClientService()
+        const clients = await cService.getClients()
 
-      res.send(200, clients)
+        res.send(200, clients)
+      } else {
+        res.send(401, 'Only admins of the API are able to use this endpoint.')
+      }
       return next()
     } catch (err) {
       console.error(err)
@@ -25,10 +31,15 @@ const main = server => {
   // Create a new Client.
   server.post({ path: '/clients', version: '1' }, async (req, res, next) => {
     try {
-      const cService = Mongo.getClientService()
-      const newClient = await cService.createClient(req.body)
+      const isAdmin = await verifyTokenIsAdmin(req.headers['request-token'])
+      if (isAdmin) {
+        const cService = Mongo.getClientService()
+        const newClient = await cService.createClient(req.body)
 
-      res.send(200, newClient)
+        res.send(200, newClient)
+      } else {
+        res.send(401, 'Only admins of the API are able to use this endpoint.')
+      }
       return next()
     } catch (err) {
       console.error(err)
@@ -40,10 +51,15 @@ const main = server => {
   // Retrieve an existing Client by ID.
   server.get({ path: '/clients/:clientId', version: '1' }, async (req, res, next) => {
     try {
-      const cService = Mongo.getClientService()
-      const client = await cService.getClient(req.params.clientId)
+      const isAdmin = await verifyTokenIsAdmin(req.headers['request-token'])
+      if (isAdmin) {
+        const cService = Mongo.getClientService()
+        const client = await cService.getClient(req.params.clientId)
 
-      res.send(200, client)
+        res.send(200, client)
+      } else {
+        res.send(401, 'Only admins of the API are able to use this endpoint.')
+      }
       return next()
     } catch (err) {
       console.error(err)
@@ -55,10 +71,15 @@ const main = server => {
   // Update an existing Client by ID.
   server.put({ path: '/clients/:clientId', version: '1' }, async (req, res, next) => {
     try {
-      const cService = Mongo.getClientService()
-      const updatedClient = await cService.updateClient(req.params.clientId, req.body)
+      const isAdmin = await verifyTokenIsAdmin(req.headers['request-token'])
+      if (isAdmin) {
+        const cService = Mongo.getClientService()
+        const updatedClient = await cService.updateClient(req.params.clientId, req.body)
 
-      res.send(200, updatedClient)
+        res.send(200, updatedClient)
+      } else {
+        res.send(401, 'Only admins of the API are able to use this endpoint.')
+      }
       return next()
     } catch (err) {
       console.error(err)
@@ -70,10 +91,15 @@ const main = server => {
   // Delete an existing Client by ID.
   server.del({ path: '/clients/:clientId', version: '1' }, async (req, res, next) => {
     try {
-      const cService = Mongo.getClientService()
-      const deletedClient = await cService.deleteClient(req.params.clientId)
+      const isAdmin = await verifyTokenIsAdmin(req.headers['request-token'])
+      if (isAdmin) {
+        const cService = Mongo.getClientService()
+        const deletedClient = await cService.deleteClient(req.params.clientId)
 
-      res.send(200, deletedClient)
+        res.send(200, deletedClient)
+      } else {
+        res.send(401, 'Only admins of the API are able to use this endpoint.')
+      }
       return next()
     } catch (err) {
       console.error(err)
@@ -85,10 +111,15 @@ const main = server => {
   // Retrieve the token for the given Client.
   server.get({ path: '/clients/:clientId/token', version: '1' }, async (req, res, next) => {
     try {
-      const cService = Mongo.getClientService()
-      const client = await cService.getClient(req.params.clientId)
+      const isAdmin = await verifyTokenIsAdmin(req.headers['request-token'])
+      if (isAdmin) {
+        const cService = Mongo.getClientService()
+        const client = await cService.getClient(req.params.clientId)
 
-      res.send(200, client.token)
+        res.send(200, client.token)
+      } else {
+        res.send(401, 'Only admins of the API are able to use this endpoint.')
+      }
       return next()
     } catch (err) {
       console.error(err)
@@ -100,10 +131,15 @@ const main = server => {
   // Update the token for the given Client.
   server.put({ path: '/clients/:clientId/token', version: '1' }, async (req, res, next) => {
     try {
-      const cService = Mongo.getClientService()
-      const updatedClient = await cService.updateClientToken(req.params.clientId, req.body)
+      const isAdmin = await verifyTokenIsAdmin(req.headers['request-token'])
+      if (isAdmin) {
+        const cService = Mongo.getClientService()
+        const updatedClient = await cService.updateClientToken(req.params.clientId, req.body)
 
-      res.send(200, updatedClient)
+        res.send(200, updatedClient)
+      } else {
+        res.send(401, 'Only admins of the API are able to use this endpoint.')
+      }
       return next()
     } catch (err) {
       console.error(err)
